requirements.txt
fastapi
uvicorn
pydantic
pandas
numpy
scikit-learn
joblib
openpyxl
xlrd
streamlit
requests
matplotlib
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-groq
faiss-cpu
sentence-transformers
huggingface-hub
python-dotenv
